menu refactor final
