+++ app/src/main/java/edu/agh/susgame/front/ui/component/menu/components/lobby/elements/LobbyComp.kt
@@ -1,4 +1,4 @@
-package edu.agh.susgame.front.ui.component.menu
+package edu.agh.susgame.front.ui.component.menu.components.lobby.elements
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -10,7 +10,6 @@ import androidx.compose.material3.Button
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -20,7 +19,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.navigation.NavController
 import edu.agh.susgame.front.model.PlayerNickname
 import edu.agh.susgame.front.model.game.Lobby
-import edu.agh.susgame.front.model.game.LobbyId
 import edu.agh.susgame.front.navigation.MenuRoute
 import edu.agh.susgame.front.service.interfaces.GameService
 import edu.agh.susgame.front.service.interfaces.LobbyService
@@ -32,7 +30,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 @Composable
-private fun LobbyContentComponent(
+internal fun LobbyComp(
     lobbyInitialState: Lobby,
     lobbyService: LobbyService,
     gameService: GameService,
@@ -162,56 +160,6 @@ private fun LobbyContentComponent(
                         text = if (isJoinButtonLoading) Translation.Button.LOADING
                         else Translation.Button.JOIN
                     )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun LobbyView(
-    lobbyId: LobbyId,
-    lobbyService: LobbyService,
-    gameService: GameService,
-    navController: NavController,
-) {
-    var lobby by remember { mutableStateOf<Lobby?>(null) }
-    var isLoading by remember { mutableStateOf(true) }
-
-    LaunchedEffect(Unit) {
-        lobbyService.getById(lobbyId)
-            .thenAccept {
-                lobby = it
-                isLoading = false
-            }
-    }
-
-    Column {
-        if (isLoading) {
-            Text(text = "${Translation.Button.LOADING}...")
-        } else {
-            when (lobby) {
-                null -> {
-                    Column {
-                        Text(text = Translation.Error.failedToLoadGame(lobbyId))
-
-                        Button(onClick = {
-                            navController.navigate(MenuRoute.SearchLobby.route)
-                        }) {
-                            Text(text = Translation.Button.GO_BACK)
-                        }
-                    }
-                }
-
-                else -> {
-                    lobby?.let {
-                        LobbyContentComponent(
-                            it,
-                            lobbyService,
-                            gameService,
-                            navController,
-                        )
-                    }
                 }
             }
         }
